@@ -1,11 +1,12 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 
 namespace AbpCompanyName.AbpProjectName.Configuration
 {
     public static class HostingEnvironmentExtensions
     {
-        public static IConfigurationRoot GetAppConfiguration(this IHostingEnvironment env)
+        public static IConfigurationRoot GetAppConfiguration(this IWebHostEnvironment env)
         {
             return AppConfigurations.Get(env.ContentRootPath, env.EnvironmentName, env.IsDevelopment());
         }
