@@ -10,10 +10,10 @@ namespace AbpCompanyName.AbpProjectName.Web.Host.Startup
        typeof(AbpProjectNameWebCoreModule))]
     public class AbpProjectNameWebHostModule: AbpModule
     {
-        private readonly IHostingEnvironment _env;
+        private readonly IWebHostEnvironment _env;
         private readonly IConfigurationRoot _appConfiguration;
 
-        public AbpProjectNameWebHostModule(IHostingEnvironment env)
+        public AbpProjectNameWebHostModule(IWebHostEnvironment env)
         {
             _env = env;
             _appConfiguration = env.GetAppConfiguration();
