@@ -5,15 +5,16 @@ using Microsoft.AspNetCore.Mvc.Razor;
 using Abp.Collections.Extensions;
 using Abp.Extensions;
 using Abp.Timing;
+using Microsoft.Extensions.Hosting;
 
 namespace AbpCompanyName.AbpProjectName.Web.Resources
 {
     public class WebResourceManager : IWebResourceManager
     {
-        private readonly IHostingEnvironment _environment;
+        private readonly IWebHostEnvironment _environment;
         private readonly List<string> _scriptUrls;
 
-        public WebResourceManager(IHostingEnvironment environment)
+        public WebResourceManager(IWebHostEnvironment environment)
         {
             _environment = environment;
             _scriptUrls = new List<string>();
